--- apps/marginfi-v2-trading/src/components/common/ActionComplete/index.ts
@@ -1,1 +1,0 @@
-export * from "./ActionComplete";
