--- apps/marginfi-v2-trading/src/components/common/ActionComplete/screens/index.ts
@@ -1,3 +1,0 @@
-export * from "./LendingScreen";
-export * from "./TradingScreen";
-export * from "./ClosePositionScreen";
