+++ apps/marginfi-v2-trading/src/components/common/Portfolio/components/close-position/close-position.tsx
@@ -1,13 +1,13 @@
 import React from "react";
+import { v4 as uuidv4 } from "uuid";
 
 import { IconX } from "@tabler/icons-react";
 import {
-  capture,
   ClosePositionActionTxns,
   ExecuteClosePositionActionProps,
+  ExecuteClosePositionAction,
   extractErrorString,
-  IndividualFlowError,
-  MultiStepToastHandle,
+
 } from "@mrgnlabs/mrgn-utils";
 import { ActiveBankInfo } from "@mrgnlabs/marginfi-v2-ui-state";
 
@@ -22,8 +22,9 @@ import { useLeveragedPositionDetails } from "~/hooks/arenaHooks";
 import { usePositionsData } from "~/hooks/usePositionsData";
 
 import { ClosePositionDialog } from "./components/close-position-dialog";
-import { handleExecuteClosePositionAction, simulateClosePosition } from "./utils/close-position-utils";
+import {  simulateClosePosition } from "./utils/close-position-utils";
 import { PreviousTxn } from "~/types";
+import { MultiStepToastController, toastManager } from "@mrgnlabs/mrgn-toasts";
 
 interface ClosePositionProps {
   arenaPool: ArenaPoolV2Extended;
@@ -35,7 +36,7 @@ interface ClosePositionProps {
 export const ClosePosition = ({ arenaPool, positionsByGroupPk, depositBanks, borrowBank }: ClosePositionProps) => {
   const [isOpen, setIsOpen] = React.useState(false);
   const [isLoading, setIsLoading] = React.useState(false);
-  const [multiStepToast, setMultiStepToast] = React.useState<MultiStepToastHandle | null>(null);
+  const [multiStepToast, setMultiStepToast] = React.useState<MultiStepToastController | null>(null);
   const [actionTxns, setActionTxns] = React.useState<ClosePositionActionTxns | null>(null);
 
   const client = useArenaClient({ groupPk: arenaPool.groupPk });
@@ -46,13 +47,11 @@ export const ClosePosition = ({ arenaPool, positionsByGroupPk, depositBanks, bor
   });
   const { connection } = useConnection();
   const { wallet } = useWallet();
-  const [platformFeeBps, broadcastType, priorityFees, setIsActionComplete, setPreviousTxn, jupiterOptions] = useUiStore(
+  const [platformFeeBps, broadcastType, priorityFees, jupiterOptions] = useUiStore(
     (state) => [
       state.platformFeeBps,
       state.broadcastType,
       state.priorityFees,
-      state.setIsActionComplete,
-      state.setPreviousTxn,
       state.jupiterOptions,
     ]
   );
@@ -66,7 +65,7 @@ export const ClosePosition = ({ arenaPool, positionsByGroupPk, depositBanks, bor
     if (!wrappedAccount || !connection) {
       return;
     }
-    const multiStepToast = new MultiStepToastHandle("Closing position", [
+    const multiStepToast = toastManager.createMultiStepToast("Closing position", [
       { label: "Loading details..." },
       { label: "Signing transaction" },
       {
@@ -93,7 +92,7 @@ export const ClosePosition = ({ arenaPool, positionsByGroupPk, depositBanks, bor
         return;
       }
 
-      multiStepToast.setSuccessAndNext();
+      multiStepToast.successAndNext();
       multiStepToast.pause();
 
       setActionTxns(actionTxns);
@@ -132,152 +131,34 @@ export const ClosePosition = ({ arenaPool, positionsByGroupPk, depositBanks, bor
   ////////////////////////////
   // Close Position Actions //
   ////////////////////////////
-  const executeAction = async (
-    params: ExecuteClosePositionActionProps,
-    arenaPool: ArenaPoolV2Extended,
-    callbacks: {
-      captureEvent: (event: string, properties?: Record<string, any>) => void;
-      setIsLoading: (loading: boolean) => void;
-      retryCallback: (txns: ClosePositionActionTxns, multiStepToast: MultiStepToastHandle) => void;
-      handleOnComplete: (txnSigs: string[]) => void;
-    }
-  ) => {
-    const action = async (params: any) => {
-      handleExecuteClosePositionAction({
-        params,
-        arenaPool,
-        captureEvent: (event, properties) => {
-          callbacks.captureEvent && callbacks.captureEvent(event, properties);
-        },
-        setIsLoading: callbacks.setIsLoading,
-        setIsComplete: callbacks.handleOnComplete,
-
-        setError: (error: IndividualFlowError) => {
-          const toast = error.multiStepToast as MultiStepToastHandle;
-          const txs = error.actionTxns as ClosePositionActionTxns;
-          const errorMessage = error.message;
-          let retry = undefined;
-          if (error.retry && toast && txs) {
-            retry = () => callbacks.retryCallback(txs, toast);
-          }
-          toast.setFailed(errorMessage, retry);
-          callbacks.setIsLoading(false);
-        },
-      });
-    };
-
-    await action(params);
-  };
-
-  const retryClosePositionAction = React.useCallback(
-    async (
-      params: ExecuteClosePositionActionProps,
-      pnl: number,
-      entryPrice: number,
-      arenaPool: ArenaPoolV2Extended
-    ) => {
-      executeAction(params, arenaPool, {
-        captureEvent: capture,
-        setIsLoading: setIsLoading,
-        retryCallback: (txns: ClosePositionActionTxns, multiStepToast: MultiStepToastHandle) => {
-          retryClosePositionAction({ ...params, actionTxns: txns, multiStepToast }, pnl, entryPrice, arenaPool);
-        },
-        handleOnComplete: (txnSigs: string[]) => {
-          setIsActionComplete(true);
-          setPreviousTxn({
-            txnType: "CLOSE_POSITION",
-            txn: txnSigs[txnSigs.length - 1],
-            positionClosedOptions: {
-              tokenBank: arenaPool.tokenBank,
-              size: positionSizeUsd,
-              leverage: Number(leverage),
-              entryPrice: entryPrice,
-              exitPrice: arenaPool.tokenBank.info.oraclePrice.priceRealtime.price.toNumber(),
-              pnl: pnl,
-              pool: arenaPool,
-            },
-          });
-          refreshGroup({
-            groupPk: arenaPool.groupPk,
-            banks: [arenaPool.tokenBank.address, arenaPool.quoteBank.address],
-            connection,
-            wallet,
-          });
-          setIsOpen(false);
-          console.log("handleOnComplete in close-position.tsx");
-        },
-      });
-    },
-    [setIsActionComplete, setPreviousTxn, positionSizeUsd, leverage, refreshGroup, connection, wallet]
-  );
-
   const handleClosePositionAction = React.useCallback(async () => {
     if (!actionTxns || !client || !multiStepToast || !arenaPool) {
       return;
     }
-    const pnl = positionData && positionData.pnl ? positionData.pnl : 0;
-    const entryPrice = positionData && positionData.entryPrice ? positionData.entryPrice : 0;
-
-    const params = {
+    
+    const props: ExecuteClosePositionActionProps = {
+      actionTxns,
+      attemptUuid: uuidv4(),
       marginfiClient: client,
-      actionTxns: actionTxns,
-      processOpts: {
-        broadcastType: broadcastType,
-        ...priorityFees,
+      multiStepToast ,
+      infoProps: {
+        token: arenaPool.tokenBank.meta.tokenSymbol,
+        tokenSize: positionSizeUsd.toString(),
+        quoteSize: positionSizeUsd.toString(),
       },
+      callbacks: {
+        captureEvent: (event: string, properties?: Record<string, any>) => {
+          console.log("captureEvent", event, properties);
+        },
+      },
+      processOpts: { ...priorityFees, broadcastType },
       txOpts: {},
-      multiStepToast: multiStepToast,
-      marginfiAccount: wrappedAccount,
-    } as ExecuteClosePositionActionProps;
 
-    return await executeAction(params, arenaPool, {
-      captureEvent: capture,
-      setIsLoading: setIsLoading,
-      handleOnComplete: (txnSigs: string[]) => {
-        setIsActionComplete(true);
-        setPreviousTxn({
-          txnType: "CLOSE_POSITION",
-          txn: txnSigs[txnSigs.length - 1],
-          positionClosedOptions: {
-            tokenBank: arenaPool.tokenBank,
-            size: positionSizeUsd,
-            leverage: Number(leverage),
-            entryPrice: entryPrice,
-            exitPrice: arenaPool.tokenBank.info.oraclePrice.priceRealtime.price.toNumber(),
-            pnl: pnl,
-            pool: arenaPool,
-          },
-        });
-        refreshGroup({
-          groupPk: arenaPool.groupPk,
-          banks: [arenaPool.tokenBank.address, arenaPool.quoteBank.address],
-          connection,
-          wallet,
-        });
-        setIsOpen(false);
-      },
-      retryCallback: (txns: ClosePositionActionTxns, multiStepToast: MultiStepToastHandle) => {
-        retryClosePositionAction({ ...params, actionTxns: txns, multiStepToast }, pnl, entryPrice, arenaPool);
-      },
-    });
-  }, [
-    actionTxns,
-    client,
-    multiStepToast,
-    arenaPool,
-    positionData,
-    broadcastType,
-    priorityFees,
-    wrappedAccount,
-    setIsActionComplete,
-    setPreviousTxn,
-    positionSizeUsd,
-    leverage,
-    refreshGroup,
-    connection,
-    wallet,
-    retryClosePositionAction,
-  ]);
+    };
+
+    ExecuteClosePositionAction(props);
+
+  }, [actionTxns, client, multiStepToast, arenaPool, positionSizeUsd, priorityFees, broadcastType]);
 
   return (
     <>
