+++ apps/marginfi-v2-trading/src/components/common/Yield/YieldCard.tsx
@@ -189,7 +189,6 @@ const YieldItem = ({
         selectedAccount={wrappedAccount}
         connected={connected}
         accountSummaryArg={accountSummary}
-        showActionComplete={false}
         hidePoolStats={["type"]}
         setDisplaySettings={setDisplaySettings}
       >
