+++ apps/marginfi-v2-trading/src/components/common/admin/admin-pool-detail/pool-detail-card/pool-detail-card.tsx
@@ -386,7 +386,6 @@ const YieldItem = ({
         selectedAccount={wrappedAccount}
         connected={connected}
         accountSummaryArg={accountSummary ?? undefined}
-        showActionComplete={false}
         hidePoolStats={["type"]}
         setDisplaySettings={setDisplaySettings}
       >
