+++ packages/mrgn-common/src/modules/transactions/index.ts
@@ -1,3 +1,2 @@
 export * from "./transaction.utils";
 export * from "./transaction.types";
-export * from "./transaction.types.v2";
