+++ packages/mrgn-ui/src/components/action-box-v2/store/action-box-store.ts
@@ -6,12 +6,8 @@ interface ActionBoxState {
   // State
   isSettingsDialogOpen: boolean;
   platformFeeBps: number;
-  isActionComplete: boolean;
-  previousTxn: PreviousTxn | null;
 
   // Actions
-  setIsActionComplete: (isActionSuccess: boolean) => void;
-  setPreviousTxn: (previousTxn: PreviousTxn) => void;
   setIsSettingsDialogOpen: (isOpen: boolean) => void;
   setPlatformFeeBps: (platformFeeBps: number) => void;
 }
@@ -29,13 +25,9 @@ const stateCreator: StateCreator<ActionBoxState, [], []> = (set, get) => ({
   isSettingsDialogOpen: false,
   priorityFee: 0,
   platformFeeBps: 30,
-  isActionComplete: false,
-  previousTxn: null,
 
   // Actions
   setIsSettingsDialogOpen: (isOpen: boolean) => set({ isSettingsDialogOpen: isOpen }),
-  setIsActionComplete: (isActionSuccess: boolean) => set({ isActionComplete: isActionSuccess }),
-  setPreviousTxn: (previousTxn: PreviousTxn) => set({ previousTxn: previousTxn }),
   setPlatformFeeBps: (platformFeeBps: number) => set({ platformFeeBps: platformFeeBps }),
 });
