--- packages/mrgn-ui/src/components/action-complete/index.tsx
@@ -1,1 +1,0 @@
-export * from "./action-complete";
