--- packages/mrgn-ui/src/components/action-complete/screens/index.ts
@@ -1,7 +1,0 @@
-export * from "./lending-screen";
-export * from "./trading-screen";
-export * from "./close-position-screen";
-export * from "./loop-screen";
-export * from "./stake-screen";
-export * from "./deposit-swap-screen";
-export * from "./repay-screen";
