+++ packages/mrgn-ui/src/components/index.ts
@@ -1,7 +1,6 @@
 export * from "./action-box-v2";
 export * from "./wallet-v2";
 export * from "./LSTDialog";
-export * from "./action-complete";
 export * from "./action-message";
 export * from "./settings";
 export * from "./stake-calculator";
